Site rebuild (#1)
* implemented all db functions

* edit and create now have tab interface

* fixed step/ing delete bug on certain edit actions, implemented recipesteps!

* first iteration of tabworld

* massive css update, nearly all funky stuff with sections fixed

* changed button padding

* added color variables, minor changes to form and recipe display pages

* form for update and create validation is done !!!!!!

* all themes finalized

* weird error fix

* tweaked matcha theme

* fixed button position

* theme switcher is a go!

* updated theme colors slightly

+++ src/lib/server/database.js
@@ -8,16 +8,56 @@ String.prototype.toTitleCase = function () {
 // PRISMA FUNCTIONS
 export async function addRecipe(recipe) {
     const recipeobject = JSON.parse(recipe);
+    let sectionlist = [];
 
-    await db.recipe.create({ 
+    var temp = await db.recipe.create({ 
         data: {
             name: recipeobject.name.toLowerCase(),
             user: recipeobject.user.toLowerCase(),
             notes: recipeobject.notes.toLowerCase(),
-            ingredients: {create: recipeobject.ingredients},
-            steps: {create: recipeobject.steps},
+            recsteps: {create: recipeobject.recsteps},
             userId: recipeobject.userid,
-        }})
+        },
+        include: {
+            sections: true,
+        },
+    })
+
+    for (const section of recipeobject.sections)
+    {        
+        var sobject = await db.section.create({
+            data: {
+                number: section.number,
+                title: section.title.toLowerCase(),
+                ingredients: {create: section.ingredients},
+                steps: {create: section.steps},
+                recipeId: temp.id,
+            },
+            include: {
+                ingredients: true,
+                steps: true,
+            },
+        })
+        sectionlist.push({id: sobject.id});
+    }
+
+    temp = await db.recipe.update({
+        where: {id: temp.id},
+        data: {
+            sections: {
+                connect: sectionlist,
+            },
+        },
+        include: {
+            sections: {
+               include: {
+                    ingredients: true,
+                    steps: true,
+               },
+            },
+            recsteps: true,
+        },
+    })
 }
 
 export async function claimRecipe(userid, recipeid) {
@@ -47,8 +87,10 @@ export async function createUser(user)
 export async function deleteAll() {
     await db.ingredient.deleteMany({});
     await db.step.deleteMany({});
+    await db.section.deleteMany({});
+    await db.recStep.deleteMany({});
     await db.recipe.deleteMany({});
-    await db.siteUser.deleteMany({});
+    // await db.siteUser.deleteMany({});
 }
 
 export async function deleteUsers() {
@@ -56,27 +98,40 @@ export async function deleteUsers() {
 }
 
 export async function deleteRecipe(id) {
-    const delId = await db.recipe.select({
+    const recipeobject = await getRecipe(id);
+
+    for (const section in recipeobject.sections)
+    {
+        if (section.id !== undefined) {
+            await db.ingredient.deleteMany({
+                where: {
+                    sectionId: section.id,
+                },
+            })
+            await db.step.deleteMany({
+                where: {
+                    sectionId: section.id,
+                },
+            })
+        }
+    }
+
+    await db.section.deleteMany({
         where: {
-            id: id,
+            recipeId: parseInt(id),
         },
     })
+
+    await db.recStep.deleteMany({
+        where: {
+            recipeId: parseInt(id),
+        },
+    })
+
 
     await db.recipe.delete({
         where: {
-            id: delId,
-        },
-    })
-
-    await db.ingredient.deleteMany({
-        where: {
-            recipeId: delId,
-        },
-    })
-
-    await db.step.deleteMany({
-        where: {
-            recipeId: delId,
+            id: parseInt(id),
         },
     })
 }
@@ -118,14 +173,37 @@ export async function getUserRecipes(arg) {
 
 export async function updateRecipe(arg) {
     const recipeobject = JSON.parse(arg);
-    for (const ingredient of recipeobject.ingredients) {
-        delete ingredient.id;
-        delete ingredient.recipeId;
+    for (const section of recipeobject.sections)
+        {   
+            if (section.id !== undefined) {
+
+                await db.ingredient.deleteMany({
+                    where: {
+                        sectionId: section.id,
+                    },
+                })
+            
+                await db.step.deleteMany({
+                    where: {
+                        sectionId: section.id,
+                    },
+                })
+        
+                await db.section.deleteMany({
+                    where: {
+                        recipeId: recipeobject.id,
+                    },
+                })
+            }
+        }
+    for (const recstep of recipeobject.recsteps)
+    {
+        delete recstep.id;
+        delete recstep.recipeId;
     }
-    for (const step of recipeobject.steps) {
-        delete step.id;
-        delete step.recipeId;
-    }
+
+    let sectionlist = [];
+
     await db.recipe.update({ 
         where: {
             id: recipeobject.id,
@@ -134,18 +212,65 @@ export async function updateRecipe(arg) {
             name: recipeobject.name.toLowerCase(),
             user: recipeobject.user.toLowerCase(),
             notes: recipeobject.notes.toLowerCase(),
-            ingredients: {
+            recsteps: {
                 deleteMany:{},
-                create: 
-                    recipeobject.ingredients,
-            },
-            steps: {
-                deleteMany:{},
-                create: 
-                    recipeobject.steps,
+                create: recipeobject.recsteps,
             },
             userId: Number(recipeobject.userid),
-        }})
+        },
+        include: {
+            sections: true,
+        },
+    })
+
+    for (const section of recipeobject.sections)
+    {   
+        for (const i of section.ingredients)
+        {
+            delete i.id;
+            delete i.sectionId;
+        }
+    
+        for (const s of section.steps)
+        {
+            delete s.id;
+            delete s.sectionId;
+        }
+
+        var sobject = await db.section.create({
+            data: {
+                number: section.number,
+                title: section.title.toLowerCase(),
+                ingredients: {create: section.ingredients},
+                steps: {create: section.steps},
+                recipeId: recipeobject.id,
+                },
+            include: {
+                ingredients: true,
+                steps: true,
+            },
+        })
+
+        sectionlist.push({id: sobject.id});
+    }
+
+
+    await db.recipe.update({
+        where: {id: recipeobject.id},
+        data: {
+            sections: {
+                connect: sectionlist,
+            },
+        },
+        include: {
+            sections: {
+               include: {
+                    ingredients: true,
+                    steps: true,
+               },
+            },
+        },
+    })
 }
 
 export async function searchName(arg) {
@@ -180,8 +305,17 @@ export async function searchIngredient(arg) {
     })
 
     for (const i of ingredients) {
-        const r = await getRecipe(i.recipeId);
-        recipes.push(r);
+        const sections = await db.section.findMany({
+            where: {
+                id: i.sectionId,
+            },
+        })
+
+        for (const s of sections)
+        {
+            const r = await getRecipe(s.recipeId);
+            recipes.push(r);
+        }
     }
     return recipes;
 }
@@ -199,11 +333,16 @@ export async function listAllUsers() {
 export async function getRecipe(id) {
     const recipe = await db.recipe.findUnique({
         where: {
-            id: Number(id),
+            id: parseInt(id),
         },
         include: {
-            ingredients: true,
-            steps: true,
+            sections: {
+               include: {
+                    ingredients: true,
+                    steps: true,
+               },
+            },
+            recsteps: true,
         },
     })
     return recipe;
